feat: add funtion to manage the state for favorite button
WHY: The meal details screen has a favorite button. if state was manage
there it won't be useful for the favorite item list screen and we have
to manage state for both of this screen and it will create a
duplication.

HOW: By moving the state in the bottom tab bar we can use it for
both of them. This is general practice in flutter where we move the
state to the parent so we can share with the childs. This process called
as a state up lefting. In future we can use the state management library
to make thing easy.