feat: add snakbar to the bottom when item is add or remove
why: when the item is add into the favorite list or item is remove from
the favorite list it is not showing any visual confirmation to the user.

how: add the _showInfoMessage method and use with condition logic which
check the item is add or remove and it is call for each step and give
visual hint to the user by using the snakbar at the bottom with the
message.